rename complytime command throughout codebase
