+++ cmd/complyctl/cli/info_test.go
