+++ cmd/complyctl/main.go
@@ -8,15 +8,15 @@ import (
 	"os"
 	"os/signal"
 
-	"github.com/complytime/complytime/cmd/complytime/cli"
+	"github.com/complytime/complyctl/cmd/complyctl/cli"
 )
 
 func main() {
 	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt)
 	defer cancel()
-	complytime := cli.New()
-	if err := complytime.ExecuteContext(ctx); err != nil {
-		cli.Error(fmt.Sprintf("error running complytime: %v", err))
+	complyctl := cli.New()
+	if err := complyctl.ExecuteContext(ctx); err != nil {
+		cli.Error(fmt.Sprintf("error running complyctl: %v", err))
 		os.Exit(1)
 	}
 }
